feat(dashboard): wallet layout split tabs
