+++ apps/dashboard/src/components/billing/Invoices/InvoicesTableActions.tsx
@@ -14,15 +14,15 @@ import {
   AlertDialogHeader,
   AlertDialogTitle,
   AlertDialogTrigger,
-} from '../ui/alert-dialog'
-import { Button } from '../ui/button'
+} from '../../ui/alert-dialog'
+import { Button } from '../../ui/button'
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuSeparator,
   DropdownMenuTrigger,
-} from '../ui/dropdown-menu'
+} from '../../ui/dropdown-menu'
 import { InvoicesTableActionsProps } from './types'
 
 export function InvoicesTableActions({ invoice, onView, onVoid, onPay }: InvoicesTableActionsProps) {
